Match entire namespace for private constants (#136)
If a constant named "Bar" is marked private, we want to also consider
any references to constants inside its namespace as private too. But
when checking for nested references, we don't want to consider
constants like "Barbie::Foo" as violations since "Barbie" is
completely unrelated to "Bar".

This commit prevents this problem by matching on the namespace
delimiter ("::") in addition to the constant.