Use cross to build linux artifact
By using cross, the build of the artifact will have a more stable ABI
because the docker image has a [well defined version of libraries][1]
that are defined, meaning that updates to "ubuntu-latest" won't impact
the compatible version of the built binaries.

This change defines the "toolchain" for rust, meaning that new versions
of rust won't automatically be picked up (unless changed to stable).

This also removes the `action-rs` uses, because they are deprecated and
the github actions install rustup by default

  [1]: https://github.com/cross-rs/cross?tab=readme-ov-file#supported-targets